synch

+++ legacy_lib/include/Log_Det.h
@@ -25,7 +25,7 @@ class Log_Det
   Log_Det(const Dynamical_Matrix& matrix, int order, double lam_max, double lam_min, bool verbose = false);
   ~Log_Det(){}
 
-  double calculate_log_det(int seed, int num_samples, bool has_zero_eigenvalue, double& variance);  
+  double calculate_log_det(long seed, int num_samples, bool has_zero_eigenvalue, double& variance);  
 
   void set_verbose(bool b) { verbose_ = b;}
   void set_debug(bool b)   { debug_   = b;}
